# openapi/polaris-mail.yaml
openapi: 3.1.0
info:
  title: polaris-mail API
  version: 0.2.0
  description: |
    HMAC-authenticated REST surface for polaris-mail. All requests are signed with
    `X-Polaris-Sig: <hex(HMAC-SHA256(secret, canonical))>` where canonical is
    `polaris-api\n<METHOD>\n<path>\n<canonical-query>\n<X-Polaris-Ts>\n<X-Polaris-Nonce>\n<sha256-hex(body)>`.

    Webhooks delivered to subscribers (the WebhookEnvelope schema below) are signed
    `X-Polaris-Sig: <hex>` over the canonical string
    `polaris-webhook\nPOST\n<path>\n<canonical-query>\n<ts>\n<nonce>\n<sha256-hex(body)>`.
    Consumers should verify with the SDK (`@polaris-mail/sdk-node`'s `verifyWebhook`
    or `polariswebhook.VerifyWebhook` from `packages/sdk-go`).
servers:
  - url: https://polaris-mail-api.workers.dev
security:
  - HMACAuth: []
components:
  securitySchemes:
    HMACAuth:
      type: apiKey
      in: header
      name: X-Polaris-Sig
  parameters:
    KeyId:
      name: X-Polaris-Key-Id
      in: header
      required: true
      schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
    Ts:
      name: X-Polaris-Ts
      in: header
      required: true
      schema: { type: string }
    Nonce:
      name: X-Polaris-Nonce
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
...
    Mailbox:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/Ulid' }
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, nullable: true }
        default_sender_id:
          allOf:
            - { $ref: '#/components/schemas/Ulid' }
          nullable: true
        created_at: { type: string }
        updated_at: { type: string }
        disabled_at: { type: string, nullable: true }
    MailboxSender:
      type: object
      required: [id, mailbox_id, domain_id, address, default_for_mailbox, created_at]
      properties:
        id: { $ref: '#/components/schemas/Ulid' }
        mailbox_id: { $ref: '#/components/schemas/Ulid' }
        domain_id: { $ref: '#/components/schemas/Ulid' }
        address: { type: string }
        local_part: { type: string, nullable: true }
        display_name: { type: string, nullable: true }
        default_for_mailbox: { type: integer, minimum: 0, maximum: 1 }
        created_at: { type: string }
        disabled_at: { type: string, nullable: true }
...
    CreateMailboxRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 2000 }
    CreateMailboxSenderRequest:
      type: object
      required: [local_part]
      properties:
        domain_id: { $ref: '#/components/schemas/Ulid' }
        local_part: { type: string, pattern: '^[a-z0-9._+-]{1,64}$' }
        display_name: { type: string, minLength: 1, maxLength: 120 }
        default_for_mailbox: { type: boolean }
    CreateMailboxReceiverRequest:
      type: object
      required: [domain_id, address_pattern]
      properties:
        domain_id: { $ref: '#/components/schemas/Ulid' }
        priority: { type: integer, default: 100 }
        address_pattern: { type: string, minLength: 1, maxLength: 512 }
        action: { type: string, enum: [webhook, forward, drop], default: webhook }
        webhook_sub_id: { $ref: '#/components/schemas/Ulid' }
        forward_to: { type: string }
...
paths:
...
  /v1/admin/mailboxes:
    get:
      summary: List mailboxes
      parameters:
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      responses:
        '200':
          description: list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Mailbox' }
    post:
      summary: Create a mailbox
      parameters:
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMailboxRequest' }
      responses:
        '201':
          description: created
  /v1/admin/mailboxes/{id}:
    get:
      summary: Get mailbox detail (with nested senders/receivers/principals/webhook_subs)
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      responses:
        '200':
          description: detail
    patch:
      summary: Update mailbox fields
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string, nullable: true }
                default_sender_id: { type: string, nullable: true }
      responses:
        '200':
          description: updated
  /v1/admin/mailboxes/{id}/disable:
    post:
      summary: Soft-disable a mailbox
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      responses:
        '200':
          description: disabled
  /v1/admin/mailboxes/{id}/enable:
    post:
      summary: Re-enable a mailbox
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      responses:
        '200':
          description: enabled
  /v1/admin/mailboxes/{id}/senders:
    post:
      summary: Create a mailbox_sender
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMailboxSenderRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MailboxSender' }
  /v1/admin/mailboxes/{id}/senders/{senderId}:
    delete:
      summary: Soft-disable a mailbox_sender
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - in: path
          name: senderId
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      responses:
        '200':
          description: disabled
  /v1/admin/mailboxes/{id}/receivers:
    post:
      summary: Create a mailbox_receiver
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMailboxReceiverRequest' }
      responses:
        '201':
          description: created
  /v1/admin/mailboxes/{id}/receivers/{receiverId}:
    patch:
      summary: Update a mailbox_receiver
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - in: path
          name: receiverId
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priority: { type: integer }
                address_pattern: { type: string }
                action: { type: string, enum: [webhook, forward, drop] }
                webhook_sub_id: { type: string, nullable: true }
                forward_to: { type: string, nullable: true }
                enabled: { type: boolean }
      responses:
        '200':
          description: updated
    delete:
      summary: Soft-disable a mailbox_receiver
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - in: path
          name: receiverId
          required: true
          schema: { $ref: '#/components/schemas/Ulid' }
        - { $ref: '#/components/parameters/KeyId' }
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      responses:
        '200':
          description: disabled
...